+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/jsonb/JsonbDeserialization.java
@@ -3,7 +3,7 @@ package io.openliberty.sample.jakarta.jsonb;
 import jakarta.json.bind.annotation.JsonbProperty;
 
 public class JsonbDeserialization{
-	
+
 	public JsonbDeserialization(String subFirstName, ChildClass child, String subfavoriteEditor) {
 		super();
 		this.subFirstName = subFirstName;
@@ -12,12 +12,12 @@ public class JsonbDeserialization{
 	}
 
 	private String subFirstName;
-    
-    private ChildClass child;
-    
-    private SubChild subChild;
-    
-    public SubChild getSubChild() {
+
+	private ChildClass child;
+
+	private SubChild subChild;
+
+	public SubChild getSubChild() {
 		return subChild;
 	}
 
@@ -34,7 +34,7 @@ public class JsonbDeserialization{
 	}
 
 	@JsonbProperty("fav_lang1")
-    private String subfavoriteEditor;    // Diagnostic: @JsonbProperty property uniqueness in subclass, multiple properties cannot have same property names.
+	private String subfavoriteEditor;    // Diagnostic: @JsonbProperty property uniqueness in subclass, multiple properties cannot have same property names.
 
 	public String getSubfavoriteEditor() {
 		return subfavoriteEditor;
@@ -43,19 +43,19 @@ public class JsonbDeserialization{
 	public void setSubfavoriteEditor(String subfavoriteEditor) {
 		this.subfavoriteEditor = subfavoriteEditor;
 	}
-	
+
 	public ChildClass getChild() {
 		return child;
-		
+
 	}
 
 	public void setChild(ChildClass child) {
 		this.child = child;
-		
+
 	}
 
 	public static class ChildClass{
-		
+
 		public ChildClass(int age, String name) {
 			super();
 			this.age = age;
@@ -65,24 +65,24 @@ public class JsonbDeserialization{
 		private String name;
 		public int getAge() {
 			return age;
-			
+
 		}
 		public void setAge(int age) {
 			this.age = age;
-			
+
 		}
 		public String getName() {
 			return name;
-			
+
 		}
 		public void setName(String name) {
 			this.name = name;
-			
+
 		}
 	}
-	
+
 	public class SubChild{
-		
+
 		private int token;
 		public int getToken() {
 			return token;
